requirements.txt
opencv_python==4.7.0.72
python-dotenv==1.0.0
